General Interface for Bayesian TidyModels

bayesian() is a way to generate a specification of a model before fitting and allows the model to be created using Stan via brms package in R. The main arguments for the model are:

prior One or more brmsprior objects created by set_prior or related functions and combined using the c method or the + operator. See also get_prior for more help.

inits Either "random" or "0". If inits is "random" (the default), Stan will randomly generate initial values for parameters. If it is "0", all parameters are initialized to zero. This option is sometimes useful for certain families, as it happens that default ("random") inits cause samples to be essentially constant. Generally, setting inits = "0" is worth a try, if chains do not behave well. Alternatively, inits can be a list of lists containing the initial values, or a function (or function name) generating initial values. The latter options are mainly implemented for internal testing but are available to users if necessary. If specifying initial values using a list or a function then currently the parameter names must correspond to the names used in the generated Stan code (not the names used in R). For more details on specifying initial values you can consult the documentation of the selected backend.

chains Number of Markov chains (defaults to 4).

iter Number of total iterations per chain (including warmup; defaults to 2000).

warmup A positive integer specifying number of warmup (aka burnin) iterations. This also specifies the number of iterations used for stepsize adaptation, so warmup samples should not be used for inference. The number of warmup should not be larger than iter and the default is iter/2.

thin Thinning rate. Must be a positive integer. Set thin > 1 to save memory and computation time if iter is large.

cores Number of cores to use when executing the chains in parallel, which defaults to 1 but we recommend setting the mc.cores option to be as many processors as the hardware and RAM allow (up to the number of chains). For non-Windows OS in non-interactive R sessions, forking is used instead of PSOCK clusters.

algorithm Character string naming the estimation approach to use. Options are "sampling" for MCMC (the default), "meanfield" for variational inference with independent normal distributions, "fullrank" for variational inference with a multivariate normal distribution, or "fixed_param" for sampling from fixed parameter values. Can be set globally for the current R session via the "brms.algorithm" option (see options).

backend Character string naming the package to use as the backend for fitting the Stan model. Options are "rstan" (the default) or "cmdstanr". Can be set globally for the current R session via the "brms.backend" option (see options). Details on the rstan and cmdstanr packages are available at https://mc-stan.org/rstan/ and https://mc-stan.org/cmdstanr/, respectively.

control A named list of parameters to control the sampler's behavior. It defaults to NULL so all the default values are used. The most important control parameters are discussed in the 'Details' section below. For a comprehensive overview see stan.

future Logical; If TRUE, the future package is used for parallel execution of the chains and argument cores will be ignored. Can be set globally for the current R session via the future option. The execution type is controlled via plan (see the examples section below).

seed The seed for random number generation to make results reproducible. If NA (the default), Stan will set the seed randomly.

refresh (non-negative integer) The number of iterations between printed screen updates. If refresh = 0, only error messages will be printed.

silent Logical; If TRUE (the default), most of the informational messages of compiler and sampler are suppressed. The actual sampling progress is still printed. Set refresh = 0 to turn this off as well. If using backend = "rstan" you can also set open_progress = FALSE to prevent opening additional progress bars.

These arguments are converted to their specific names at the time that the model is fit. Other options and argument can be set using set_engine(). If left to their defaults here (NULL), the values are taken from the underlying model functions. If parameters need to be modified, update() can be used in lieu of recreating the object from scratch.

bayesian(
  mode = "regression",
  prior = NULL,
  inits = NULL,
  chains = NULL,
  iter = NULL,
  warmup = NULL,
  thin = NULL,
  cores = NULL,
  algorithm = NULL,
  backend = NULL,
  control = NULL,
  future = NULL,
  seed = NULL,
  refresh = NULL,
  silent = NULL
)

# S3 method for bayesian
update(
  object,
  parameters = NULL,
  prior = NULL,
  inits = NULL,
  chains = NULL,
  iter = NULL,
  warmup = NULL,
  thin = NULL,
  cores = NULL,
  algorithm = NULL,
  backend = NULL,
  control = NULL,
  future = NULL,
  seed = NULL,
  refresh = NULL,
  silent = NULL,
  fresh = FALSE,
  ...
)

bayesian_predict(
  object,
  newdata = NULL,
  re_formula = NULL,
  transform = NULL,
  resp = NULL,
  negative_rt = FALSE,
  nsamples = NULL,
  subset = NULL,
  sort = FALSE,
  ntrys = 5,
  cores = getOption("mc.cores", 1),
  summary = TRUE,
  robust = FALSE,
  probs = c(0.025, 0.975),
  ...
)

bayesian_read(file)

bayesian_write(x, file)

Arguments

mode
A single character string for the type of model. Possible values for this model are "unknown", "regression", or "classification".

prior
One or more brmsprior objects created by set_prior or related functions and combined using the c method or the + operator. See also get_prior for more help.

inits
Either "random" or "0". If inits is "random" (the default), Stan will randomly generate initial values for parameters. If it is "0", all parameters are initialized to zero. This option is sometimes useful for certain families, as it happens that default ("random") inits cause samples to be essentially constant. Generally, setting inits = "0" is worth a try, if chains do not behave well. Alternatively, inits can be a list of lists containing the initial values, or a function (or function name) generating initial values. The latter options are mainly implemented for internal testing but are available to users if necessary. If specifying initial values using a list or a function then currently the parameter names must correspond to the names used in the generated Stan code (not the names used in R). For more details on specifying initial values you can consult the documentation of the selected backend.

chains
Number of Markov chains (defaults to 4).

iter
Number of total iterations per chain (including warmup; defaults to 2000).

warmup
A positive integer specifying number of warmup (aka burnin) iterations. This also specifies the number of iterations used for stepsize adaptation, so warmup samples should not be used for inference. The number of warmup should not be larger than iter and the default is iter/2.

thin
Thinning rate. Must be a positive integer. Set thin > 1 to save memory and computation time if iter is large.

cores
Number of cores to use when executing the chains in parallel, which defaults to 1 but we recommend setting the mc.cores option to be as many processors as the hardware and RAM allow (up to the number of chains). For non-Windows OS in non-interactive R sessions, forking is used instead of PSOCK clusters.

algorithm
Character string naming the estimation approach to use. Options are "sampling" for MCMC (the default), "meanfield" for variational inference with independent normal distributions, "fullrank" for variational inference with a multivariate normal distribution, or "fixed_param" for sampling from fixed parameter values. Can be set globally for the current R session via the "brms.algorithm" option (see options).

backend
Character string naming the package to use as the backend for fitting the Stan model. Options are "rstan" (the default) or "cmdstanr". Can be set globally for the current R session via the "brms.backend" option (see options). Details on the rstan and cmdstanr packages are available at https://mc-stan.org/rstan/ and https://mc-stan.org/cmdstanr/, respectively.

control
A named list of parameters to control the sampler's behavior. It defaults to NULL so all the default values are used. The most important control parameters are discussed in the 'Details' section below. For a comprehensive overview see stan.

future
Logical; If TRUE, the future package is used for parallel execution of the chains and argument cores will be ignored. Can be set globally for the current R session via the future option. The execution type is controlled via plan (see the examples section below).

seed
The seed for random number generation to make results reproducible. If NA (the default), Stan will set the seed randomly.

refresh
(non-negative integer) The number of iterations between printed screen updates. If refresh = 0, only error messages will be printed.

silent
Logical; If TRUE (the default), most of the informational messages of compiler and sampler are suppressed. The actual sampling progress is still printed. Set refresh = 0 to turn this off as well. If using backend = "rstan" you can also set open_progress = FALSE to prevent opening additional progress bars.

object
A Bayesian model specification.

parameters
A 1-row tibble or named list with main parameters to update. If the individual arguments are used, these will supersede the values in parameters. Also, using engine arguments in this object will result in an error.

fresh
A logical for whether the arguments should be modified in-place of or replaced wholesale.

...
Other arguments passed to internal functions.

newdata
An optional data.frame for which to evaluate predictions. If NULL (default), the original data of the model is used. NA values within factors are interpreted as if all dummy variables of this factor are zero. This allows, for instance, to make predictions of the grand mean when using sum coding.

re_formula
formula containing group-level effects to be considered in the prediction. If NULL (default), include all group-level effects; if NA, include no group-level effects.

transform
(Deprecated) A function or a character string naming a function to be applied on the predicted responses before summary statistics are computed.

resp
Optional names of response variables. If specified, predictions are performed only for the specified response variables.

negative_rt
Only relevant for Wiener diffusion models. A flag indicating whether response times of responses on the lower boundary should be returned as negative values. This allows to distinguish responses on the upper and lower boundary. Defaults to FALSE.

nsamples
Positive integer indicating how many posterior samples should be used. If NULL (the default) all samples are used. Ignored if subset is not NULL.

subset
A numeric vector specifying the posterior samples to be used. If NULL (the default), all samples are used.

sort
Logical. Only relevant for time series models. Indicating whether to return predicted values in the original order (FALSE; default) or in the order of the time series (TRUE).

ntrys
Parameter used in rejection sampling for truncated discrete models only (defaults to 5). See Details for more information.

summary
Should summary statistics be returned instead of the raw values? Default is TRUE.

robust
If FALSE (the default) the mean is used as the measure of central tendency and the standard deviation as the measure of variability. If TRUE, the median and the median absolute deviation (MAD) are applied instead. Only used if summary is TRUE.

probs
The percentiles to be computed by the quantile function. Only used if summary is TRUE.

file
A character string of the file path to brmsfit object saved via saveRDS.

x
An object of class brmsfit, which contains the posterior samples along with many other useful information about the model. Use methods(class = "brmsfit") for an overview on available methods.

Value

An updated model specification.

An object of class brmsfit, which contains the posterior samples along with many other useful information about the model. Use methods(class = "brmsfit") for an overview on available methods.

Details

The data given to the function are not saved and are only used to determine the mode of the model. For bayesian(), the possible modes are "regression" and "classification".

The model can be created by the fit() function using the following engines:

brms: "stan"

Engine Details

Engines may have pre-set default arguments when executing the model fit call. For this type of model, the template of the fit calls are:

bayesian() %>% 
  set_engine("stan") %>% 
  translate()

## Bayesian Model Specification (regression)
## 
## Computational engine: stan 
## 
## Model fit template:
## brms::brm(formula = missing_arg(), data = missing_arg(), family = brms::brmsfamily("gaussian"))

See also

Examples

bayesian()

#> Bayesian Model Specification (regression)
#> 

show_model_info("bayesian")

#> Information for `bayesian`
#>  modes: unknown, classification, regression 
#> 
#>  engines: 
#>    classification: stan
#>    regression:     stan
#> 
#>  no registered arguments.
#> 
#>  fit modules:
#>  engine           mode
#>    stan classification
#>    stan     regression
#> 
#>  prediction modules:
#>              mode engine                                        methods
#>    classification   stan class, conf_int, pred_int, prob, quantile, raw
#>        regression   stan     conf_int, numeric, pred_int, quantile, raw
#> 

bayesian(mode = "classification")

#> Bayesian Model Specification (classification)
#> 

bayesian(mode = "regression")

#> Bayesian Model Specification (regression)
#> 

if (FALSE) {
bayesian_mod <-
  bayesian() %>%
  set_engine("stan") %>%
  fit(
    rating ~ treat + period + carry + (1 | subject),
    data = inhaler
  )

summary(bayesian_mod$fit)
}

# -------------------------------------------------------------------------
model <- bayesian(inits = "random")
model

#> Bayesian Model Specification (regression)
#> 
#> Main Arguments:
#>   inits = random
#> 

update(model, inits = "0")

#> Bayesian Model Specification (regression)
#> 
#> Main Arguments:
#>   inits = 0
#> 

update(model, inits = "0", fresh = TRUE)

#> Bayesian Model Specification (regression)
#> 
#> Main Arguments:
#>   inits = 0
#>